dev: ui completed
