+++ src/app/register/page.tsx
@@ -1,5 +1,6 @@
 "use client"
 
+import Loader from "@/components/Loader"
 import { registerUser } from "@/redux/actions"
 import { useAppDispatch } from "@/utility/type"
 import { useRouter } from "next/navigation"
@@ -14,41 +15,59 @@ const Registration = () => {
     password: "",
     confirmPassword: "",
   })
-
-  // useEffect(() => {
-  //   router.push("/login")
-  // }, [router])
+  const [loading, setLoading] = useState<Boolean>(false)
 
   const handleRegistration = () => {
     if (data?.password !== data?.confirmPassword)
       return toast.warn("Passwords don't match")
     else if (data?.password?.length < 5)
       return toast.warn("Minimum 5 characters")
-    else dispatch(registerUser(data))
+    else{
+        setLoading(true)
+       dispatch(registerUser(data))
+    }
   }
+
+  if (loading) return <Loader />;
+
+
   return (
-    <div>
-      <input
-        onChange={(e) => setData({ ...data, userName: e.target.value })}
-        type="text"
-      />
-      UserName
-      <br />
-      <input
-        onChange={(e) => setData({ ...data, password: e.target.value })}
-        type="text"
-      />
-      Password
-      <br />
-      <input
-        onChange={(e) => setData({ ...data, confirmPassword: e.target.value })}
-        type="text"
-      />
-      Confirm Password
-      <br />
-      <button onClick={handleRegistration}>Send</button>
+    <div className="h-screen flex justify-center items-center">
+      <div className="flex h-full flex-col justify-center items-center lg:w-1/4 md:w-2/5 w-4/5 space-y-6 py-4 px-6 rounded-md ">
+        <h1 className="text-center text-2xl text-black mb-6 dark:text-white">
+          REGISTER
+        </h1>
+        <div className="space-y-4 w-full flex flex-col">
+          <input
+            placeholder="Enter your username"
+            className="py-3 px-3 text-sm rounded-md outline-none bg-gray-300"
+            onChange={(e) => setData({ ...data, userName: e.target.value })}
+            type="text"
+          />
+          <input
+            placeholder="Enter your password"
+            className="py-3 px-3 text-sm rounded-md outline-none bg-gray-300"
+            onChange={(e) => setData({ ...data, password: e.target.value })}
+            type="password"
+          />
+          <input
+            placeholder="Confirm your password"
+            className="py-3 px-3 text-sm rounded-md outline-none bg-gray-300"
+            onChange={(e) =>
+              setData({ ...data, confirmPassword: e.target.value })
+            }
+            type="password"
+          />
+        </div>
+        <button
+          className="bg-blue py-2 w-full font-medium rounded-md text-white"
+          onClick={handleRegistration}
+        >
+          REGISTER
+        </button>
+      </div>
     </div>
-  )
+  );
 }
 
 export default Registration
